+++ accounts_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AccountsAppConfig(AppConfig):
+    name = 'accounts_app'
