+++ cart_app/templates/checkout/home.html
@@ -1,0 +1,10 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<h1>Checkout</h1>
+
+{{ object.shipping_total }}
+{{ object.total }}
+{{ object.status }}
+
+{% endblock content %}
