+++ cart_app/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
-from .views import cart_home, cart_udpate
+from .views import cart_home, cart_udpate, checkout_home
 
 urlpatterns = [
     path('', cart_home,name='cart_home'),
     path('update', cart_udpate,name='cart_update'),
+    path('checkout', checkout_home,name='checkout'),
 ]
